pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Aleksey Zvyagintsev",email = "alex0236889@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "django-filter (>=25.2,<26.0)",
    "black (==25.12.0)",
    "mypy (==1.19.1)",
    "docutils (==0.22.4)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "poetry-core (>=2.0.0)",
    "celery (>=5.6.2,<6.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^6.0.1"
python-dotenv = "^1.1.1"
psycopg2-binary = "==2.9.11"
poetry-core = "^2.0.0"
apscheduler = "==3.11.1"
django-apscheduler = "^0.7.0"
email-validator = "^2.3.0"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^25.11.0"
isort = "7.0.0"
mypy = "^1.14.1"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = ['venv', '.venv', 'tests', '__pycache__']
check_untyped_defs = true
strict = true
warn_unreachable = true

[tool.isort]
line_length = 119
multi_line_output = 1

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.mypy_cache
    | \.venv
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''
